+++ frontend/src/metabase/visualizer/utils/index.ts
@@ -1,0 +1,4 @@
+export * from "./column";
+export * from "./data-source";
+export * from "./drag-and-drop";
+export * from "./misc";
